#554: move xpath.cc/h to src/mxml
